web/requirements.txt
Flask==2.0.3
psycopg2-binary==2.9.5
pytest==7.2.0
python-dotenv==1.0.0